Split subscriber and publisher tests.
